requirements.txt
discord.py-self
Pillow
gofile
requests
pymongo